refactor: move task config loader to project top level
