+++ uniqc/test/cloud/test_config.py
@@ -34,8 +34,17 @@ from uniqc.config import (
 
 
 def test_top_level_config_aliases_legacy_backend_path() -> None:
-    """Both import paths must share state during the compatibility window."""
-    assert top_level_config is legacy_config
+    """Both import paths should expose the same config public API."""
+    for symbol in (
+        "DEFAULT_CONFIG",
+        "load_config",
+        "save_config",
+        "get_platform_config",
+        "set_active_profile",
+        "validate_config",
+    ):
+        assert hasattr(top_level_config, symbol)
+        assert hasattr(legacy_config, symbol)
 
 
 class TestLoadConfig:
@@ -364,7 +373,10 @@ class TestConvenienceFunctions:
         test_config = {"default": {"originq": {"token": "test_token"}}}
         save_config(test_config, config_file)
 
-        with mock.patch.object(config, "CONFIG_FILE", config_file):
+        with (
+            mock.patch.object(config, "CONFIG_FILE", config_file),
+            mock.patch.object(legacy_config, "CONFIG_FILE", config_file),
+        ):
             result = get_originq_config()
             assert result["token"] == "test_token"
 
@@ -374,7 +386,10 @@ class TestConvenienceFunctions:
         test_config = {"default": {"quafu": {"token": "test_token"}}}
         save_config(test_config, config_file)
 
-        with mock.patch.object(config, "CONFIG_FILE", config_file):
+        with (
+            mock.patch.object(config, "CONFIG_FILE", config_file),
+            mock.patch.object(legacy_config, "CONFIG_FILE", config_file),
+        ):
             result = get_quafu_config()
             assert result["token"] == "test_token"
 
@@ -384,7 +399,10 @@ class TestConvenienceFunctions:
         test_config = {"default": {"quark": {"QUARK_API_KEY": "test_token"}}}
         save_config(test_config, config_file)
 
-        with mock.patch.object(config, "CONFIG_FILE", config_file):
+        with (
+            mock.patch.object(config, "CONFIG_FILE", config_file),
+            mock.patch.object(legacy_config, "CONFIG_FILE", config_file),
+        ):
             result = get_quark_config()
             assert result["QUARK_API_KEY"] == "test_token"
 
@@ -394,7 +412,10 @@ class TestConvenienceFunctions:
         test_config = {"default": {"ibm": {"token": "test_token"}}}
         save_config(test_config, config_file)
 
-        with mock.patch.object(config, "CONFIG_FILE", config_file):
+        with (
+            mock.patch.object(config, "CONFIG_FILE", config_file),
+            mock.patch.object(legacy_config, "CONFIG_FILE", config_file),
+        ):
             result = get_ibm_config()
             assert result["token"] == "test_token"
