+++ uniqc/test/test_package_structure.py
@@ -50,7 +50,7 @@ def test_uniqc_top_level_packages_match_architecture() -> None:
 
 
 def test_common_user_imports_are_flat() -> None:
-    import uniqc.backend_adapter.config as legacy_config
+    import uniqc.backend_adapter.config as backend_adapter_config
     from uniqc import (
         BackendInfo,
         Circuit,
@@ -73,4 +73,6 @@ def test_common_user_imports_are_flat() -> None:
     assert calculate_expectation is not None
     assert qft_circuit is not None
     assert QuantumLayer is not None
-    assert config is legacy_config
+    assert config is not None
+    assert hasattr(config, "load_originq_config")
+    assert hasattr(backend_adapter_config, "get_platform_config")
